PLATUI-64: Refactored paths out into shared file
