--- application/filters/hmrc-design-system.js
@@ -1,6 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  isArray: function (obj) { return Array.isArray(obj) },
-  getDirectoryFromFilepath: function (filepath) { return path.dirname(filepath) }
-}
