+++ application/templates/partials/_example.njk
@@ -5,8 +5,8 @@
   {% set exampleId = "example-" + exampleRef %}
   {% if params.welsh %}
     {% set welshExampleRef = params.welsh | default("default") %}
-    {% set welshExamplePath = "src" + welshExampleURL + "index.njk" %}
     {% set welshExampleURL = "/design-library/" + params.item + "/" + welshExampleRef + "/" %}
+    {% set welshExamplePath = "src" + welshExampleURL + "index.njk" %}
   {% endif %}
   <div class="app-example-wrapper app-example__language-switch--en-activated" id="{{ exampleId }}" data-module="app-tabs app-language-switch-example">
     <div class="app-example app-example--tabs">
