+++ lib/globals.js
@@ -2,14 +2,8 @@ const fs = require('fs')
 const nunjucks = require('nunjucks')
 const matter = require('gray-matter')
 const beautify = require('js-beautify').html
-const pathFromRoot = require('../tasks/gulp/util').pathFromRoot
 
-const templatePaths = [
-  pathFromRoot('node_modules', 'hmrc-frontend', 'components'),
-  pathFromRoot('application', 'templates'),
-  pathFromRoot('application', 'templates', 'partials'),
-  pathFromRoot('src')
-]
+const templatePaths = require('./templatePaths')
 
 nunjucks.configure(templatePaths)
 
@@ -26,6 +20,7 @@ exports.getHTMLCode = path => {
     html = nunjucks.renderString(content)
   } catch (err) {
     if (err) {
+      console.log(err)
       console.log('Could not get HTML code from ' + path)
     }
   }
