+++ lib/templatePaths.js
@@ -1,0 +1,8 @@
+const pathFromRoot = require('../util/pathFromRoot')
+
+module.exports = [
+  pathFromRoot('node_modules', 'hmrc-frontend', 'components'),
+  pathFromRoot('application', 'templates'),
+  pathFromRoot('application', 'templates', 'partials'),
+  pathFromRoot('src'),
+]
