+++ tasks/gulp/clean.js
@@ -1,6 +1,6 @@
 const gulp = require('gulp')
 const del = require('del')
-const pathFromRoot = require('./util').pathFromRoot
+const pathFromRoot = require('../../util/pathFromRoot')
 
 gulp.task('clean', () => del.sync([
   pathFromRoot('dist', '/**/*')
