+++ tasks/gulp/serve.js
@@ -1,6 +1,7 @@
 const gulp = require('gulp')
 const connect = require('gulp-connect')
-const pathFromRoot = require('./util').pathFromRoot
+
+const pathFromRoot = require('../../util/pathFromRoot')
 
 gulp.task('serve', () => {
   connect.server({
