--- tasks/gulp/util.js
@@ -1,5 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  pathFromRoot: (...args) => path.join(process.cwd(), ...args)
-}
