+++ util/pathFromRoot.js
@@ -1,0 +1,3 @@
+const path = require('path')
+
+module.exports = (...args) => path.join(process.cwd(), ...args)
